feat(framework): moving to solid
