+++ index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link href="https://fonts.googleapis.com/css2?family=Limelight" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700" rel="stylesheet">
+    <link rel='icon' type='image/png' href='src/assets/favicon.png'>
+    <link rel='icon' type='image/svg+xml' href='src/assets/favicon.svg'>
+    <link rel="alternate icon" href="src/assets/favicon.ico">
+    <title>Hashtags - Generate Random Hashtags</title>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+
+    <script src="/src/index.jsx" type="module"></script>
+  </body>
+</html>
