+++ layouts/default.jsx
@@ -1,0 +1,10 @@
+// eslint-disable-next-line no-unused-vars
+import {Route} from 'solid-app-router';
+
+function Default() {
+  return (
+    <Route />
+  );
+}
+
+export default Default;
