+++ package.json
@@ -22,14 +22,18 @@
   "devDependencies": {
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
-    "@nuxtjs/axios": "^5.13.6",
     "eslint": "^7.29.0",
     "eslint-config-xo-space": "^0.28.0",
-    "eslint-plugin-vue": "^7.11.1",
     "husky": "^6.0.0",
-    "nuxt": "^2.15.7",
-    "nuxt-animejs": "^1.2.1",
-    "nuxt-vite": "^0.1.1",
-    "nuxt-windicss": "^1.0.6"
+    "vite": "^2.4.0",
+    "vite-plugin-solid": "^2.0.0",
+    "vite-plugin-windicss": "^1.2.3",
+    "windicss": "^3.1.4"
+  },
+  "dependencies": {
+    "animejs": "^3.2.1",
+    "axios": "^0.21.1",
+    "solid-app-router": "^0.0.50",
+    "solid-js": "^1.0.0"
   }
 }
